Lab 1, OOP Discipline, Programming language C++
Working functionality that meets the requirements of laboratory work

Problem -> Create a class in which to implement functions for working
with one-dimensional arrays: a) get the intersection of array elements
b)get the union of array elements.		Allocate memory for
arrays dynamically.Use constructor with parameters, copy constructor.
The destrcutor must free memory allocated for arrays.